+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 ### Changed
+- Portfolio Data comes from Snowflake table
+- Use Portfolio Datasource to create security and company objects
 ### Removed
 
 ## [1.1.1] - 2023-09-21